+++ src/plays/simple-live-chat/_variables.scss
@@ -14,3 +14,4 @@ $simple-header-color: #bae6fd;
 $simple-text-color: #444444;
 $simple-user-border-color: #94a3b8;
 $simple-gray-color: gray;
+$simple-default-background-color: #f6f6f9;
