+++ src/plays/simple-live-chat/components/chat/components/User.jsx
@@ -1,0 +1,20 @@
+// vendors
+import { useState } from "react";
+
+// assets
+import userIcon from "../../../images/user_icon.png";
+
+const User = ({ photoURL, onClick }) => {
+  const [defaultImage, setDefaultImage] = useState(photoURL);
+  return (
+    <img
+      src={defaultImage}
+      alt='avatar'
+      onClick={onClick}
+      className='simple-live-chat-user'
+      onError={() => setDefaultImage(userIcon)}
+    />
+  );
+};
+
+export default User;
