+++ src/plays/simple-live-chat/components/chat/hooks/useDateFormat.js
@@ -1,0 +1,11 @@
+import format from "date-fns/format";
+
+const useDateFormat = () => {
+  const dateTimeFormatter = (val, formatText = "hh:mm bbb") => {
+    return format(new Date(val * 1000), formatText);
+  };
+
+  return [dateTimeFormatter];
+};
+
+export default useDateFormat;
